// src/features/swal-qol/types/swq.ts
import { ISwalQOL } from "./swalQOL";
import { IGeneralSwalQOL } from "./swalQOL";
import { IEatingSwalQOL } from "./swalQOL";
import { IEatingTimeSwalQOL } from "./swalQOL";
import { IPhysicalIssuesSwalQOL } from "./swalQOL";
import { IFoodChoicesSwalQOL } from "./swalQOL";
import { ICommunicationSwalQOL } from "./swalQOL";
import { IConcernsSwalQOL } from "./swalQOL";
import { IEmotionsSwalQOL } from "./swalQOL";
import { ISocialLifeSwalQOL } from "./swalQOL";
import { IPhysicalSymptomsSwalQOL } from "./swalQOL";
import { ISleepIssuesSwalQOL } from "./swalQOL";
import { IScoreGlobalSwalQOL } from "./swalQOL";
import { IGeneralQuestionsSwalQOL } from "./swalQOL";
export interface ISwq {
    identificatioSwalQOL: ISwalQOL;
    generalSwalQOL: IGeneralSwalQOL;
    eatingSwalQOL: IEatingSwalQOL;
    eatingTimeSwalQOL: IEatingTimeSwalQOL;
    physicalIssuesSwalQOL: IPhysicalIssuesSwalQOL;
    foodChoicesSwalQOL: IFoodChoicesSwalQOL;
    communicationSwalQOL: ICommunicationSwalQOL;
    concernsSwalQOL: IConcernsSwalQOL;
    emotionsSwalQOL: IEmotionsSwalQOL;
    socialLifeSwalQOL: ISocialLifeSwalQOL;
    physicalSymptomsSwalQOL: IPhysicalSymptomsSwalQOL;
    sleepIssuesSwalQOL: ISleepIssuesSwalQOL;
    scoreGlobalSwalQOL: IScoreGlobalSwalQOL;
    generalQuestionsSwalQOL: IGeneralQuestionsSwalQOL;
};